requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
numpy==2.4.3
pillow==12.1.1
pycparser==3.0
pycryptodome==3.23.0
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
